[librptexture] RpGdiplusBackend: Added a (size_t) cast to prevent a warning from ALIGN_BYTES().
The constness causes problems with __typeof__() in MinGW builds:

In file included from src/librptexture/stdafx.h:43,
                 from src/librptexture/img/RpGdiplusBackend.cpp:12:
src/librptexture/img/RpGdiplusBackend.cpp: In member function 'HBITMAP__* LibRpTexture::RpGdiplusBackend::convBmpData_CI8(const BitmapData*)':
src/librpbase/common.h:140:68: warning: type qualifiers ignored on cast result type [-Wignored-qualifiers]
  140 | # define ALIGN_BYTES(a, x) (((x)+((a)-1)) & ~((__typeof__(x))((a)-1)))
      |                                                                    ^
src/librptexture/img/RpGdiplusBackend.cpp:921:26: note: in expansion of macro 'ALIGN_BYTES'
  921 |  const int hbmp_stride = ALIGN_BYTES(4, row_bytes);
      |                          ^~~~~~~~~~~

+++ src/librptexture/img/RpGdiplusBackend.cpp
@@ -918,7 +918,7 @@ HBITMAP RpGdiplusBackend::convBmpData_CI8(const Gdiplus::BitmapData *pBmpData)
 	// Copy the data from the GDI+ bitmap to the HBITMAP directly.
 	// HBITMAP stride is a multiple of 4.
 	const size_t row_bytes = pBmpData->Width;
-	const int hbmp_stride = ALIGN_BYTES(4, row_bytes);
+	const int hbmp_stride = ALIGN_BYTES(4, (size_t)row_bytes);
 	const int gdip_stride = pBmpData->Stride;
 	if (hbmp_stride == gdip_stride) {
 		// Copy the entire image all at once.
